requirements.txt
alembic==1.4.3
aniso8601==8.0.0
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coverage==5.3
coveralls==2.2.0
cryptography==3.2.1
docopt==0.6.2
flake8==3.8.4
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
nose2==0.9.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
